Update checkout page
