+++ src/pages/checkout.module.scss
@@ -1,0 +1,33 @@
+@import "../variables.scss";
+
+.cart {
+  display: inline-block;
+  vertical-align: top;
+  @media (min-width: $lg) {
+    width: 60%;
+    padding-right: 0.5rem;
+  }
+  .header {
+    font-size: 1.5rem;
+    font-weight: bold;
+    padding-bottom: 1rem;
+    border-bottom: 1px solid $grey;
+    .itemCount {
+      color: $grey;
+    }
+  }
+}
+
+.summary {
+  display: inline-block;
+  vertical-align: top;
+  @media (min-width: $lg) {
+    width: 40%;
+    padding-left: 0.5rem;
+  }
+  .header {
+    font-size: 1.5rem;
+    font-weight: bold;
+    padding-bottom: 1rem;
+  }
+}
